# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CreatorHub API",
        "description": "Documentation de l'API CreatorHub",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/offres": {
            "get": {
                "tags": [
                    "Offres"
                ],
                "summary": "Lister toutes les offres",
                "operationId": "1ad8c9cc0acc58b6150e8e26720f179e",
                "responses": {
                    "200": {
                        "description": "Liste des offres récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Offres"
                ],
                "summary": "Créer une nouvelle offre",
                "operationId": "06984a0bba650db64aedbd14e79b4736",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "budget"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Recherche Graphiste"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Création d'un logo"
                                    },
                                    "budget": {
                                        "type": "number",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre créée avec succès"
                    }
                }
            }
        },
        "/api/offres/{id}": {
            "get": {
                "tags": [
                    "Offres"
                ],
                "summary": "Afficher une offre",
                "operationId": "6896d86f6130d6d423825df0c6d2004a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'offre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre trouvée"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Offres"
                ],
                "summary": "Modifier une offre",
                "operationId": "393ef2daeb54a5d9c95acf0c98eb5b8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "title": "Recherche Développeur Laravel",
                                "description": "Développement API REST",
                                "budget": 2500
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre modifiée avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Offres"
                ],
                "summary": "Supprimer une offre",
                "operationId": "89b2f98fe2ba17929d48963fbb2c48fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'offre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre supprimée avec succès"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    }
                }
            }
        },
        "/api/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Tester Swagger",
                "operationId": "af4eef7c415174b57a22c482494a6a28",
                "responses": {
                    "200": {
                        "description": "API fonctionnelle"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Offres",
            "description": "Offres"
        },
        {
            "name": "Test",
            "description": "Test"
        }
    ]
}